:bug: Fixing VMs sometimes coming up blank
Garbage collection was removing the old VMs from docker thus making it
so we couldn't rely on the simple start and stop only. Added the
directory with the fully configured services into the database. If the
service isn't available to start/stop we will use the directory to start
it.

closes #8