Trimming Potential Bots from View
This view was grabbing all messages from people that received a ban.
This gets people that are cronically getting banned or flags a lot of
people for streams where banning people is part of the banter. A more
accurate assessment of the data would be to say we want to see someone
that got banned in a single channel after only writing 1-3 messages
(inclusive). This is going to get the one off hate watchers but if you
sort the data and compare the rows side by side you'll see the patterns
for bot messages.